+++ CHANGELOG.rst
@@ -7,6 +7,14 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+Unreleased
+----------
+
+**Fixed**
+
+* Fixed Windows OS "Display not in os.environ" issue in `connect`
+
+
 0.1.0
 ----------